Move InterpolatingTreeMap and friends to wpimath
