+++ wpimath/src/main/java/edu/wpi/first/math/interpolation/InterpolatingTreeMap.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.util;
+package edu.wpi.first.math.interpolation;
 
 import java.util.Comparator;
 import java.util.TreeMap;
